VMware Plugin: Save VM configuration and recreate VM before restore
The plugin now also saves the VM configuration on backup and recreates
the VM if it was delete before restoring the virtual disk. It is also
possible now to restore to different folder, host, cluster or datastore.
Custom attributes are saved and restored, uuid and MAC address are
restored if there are no duplicate conflicts.