libcloud-plugin: Refactor for current Python
Major change is to use threading instead of multiprocessing as that does
not work with Bareos plugins and current Pyhton versions.
Unfortunately libcloud is not fully thread safe, when libcloud calls
run into timeouts throwing exceptions, which can happen on bad network
connections, next calls get stuck within the worker threads. So the
default for the fail_on_download_error is now yes and the plugin code
tries to detect unresponsive worker threads and terminates the job.